# Ruby: uglier but faster implementation