Mostly I just replaced `.each` and `.each_with_index` with `while` loop, locally it performs almost two times faster. Looks ugly but performs better :upside_down_face: 
also performance can be improved by just enabling YJIT and using `Oj` gem instead of `JSON` from std

CPU: AMD Ryzen 7 4700U
RAM: 16 GiB
Ruby: 3.3.4

Old:
```
Running ruby
Ruby:

        Processing time (w/o IO): 2883ms
        total: 2.99s memory: 25684k
        Processing time (w/o IO): 3206ms
        total: 3.31s memory: 25492k
        Processing time (w/o IO): 2979ms
        total: 3.08s memory: 25616k
        Processing time (w/o IO): 2973ms
        total: 3.07s memory: 25756k
        Processing time (w/o IO): 3027ms
        total: 3.13s memory: 25448k
        Processing time (w/o IO): 2889ms
        total: 2.99s memory: 25872k
        Processing time (w/o IO): 2944ms
        total: 3.05s memory: 25700k
        Processing time (w/o IO): 2894ms
        total: 3.00s memory: 25816k
        Processing time (w/o IO): 2897ms
        total: 2.99s memory: 25496k
        Processing time (w/o IO): 3055ms
        total: 3.16s memory: 25876k
Checking output
Success: related_posts_ruby.json is valid!
```
New:
```
Running ruby
Ruby:

        Processing time (w/o IO): 1518ms
        total: 1.62s memory: 25532k
        Processing time (w/o IO): 1516ms
        total: 1.62s memory: 25516k
        Processing time (w/o IO): 1504ms
        total: 1.61s memory: 25476k
        Processing time (w/o IO): 1534ms
        total: 1.64s memory: 25464k
        Processing time (w/o IO): 1523ms
        total: 1.64s memory: 25328k
        Processing time (w/o IO): 1538ms
        total: 1.65s memory: 25548k
        Processing time (w/o IO): 1509ms
        total: 1.61s memory: 25536k
        Processing time (w/o IO): 1512ms
        total: 1.60s memory: 25508k
        Processing time (w/o IO): 1561ms
        total: 1.67s memory: 25376k
        Processing time (w/o IO): 1486ms
        total: 1.59s memory: 25456k
Checking output
Success: related_posts_ruby.json is valid!
```