@jinyus 2024-10-10
Main:
```sh
| Ruby | 2.29 s | 35.18 s | 316.54 s | 354.00 s |
```

This PR:
```yaml
Ruby:

        Processing time (w/o IO): 1461ms
        total: 1.62s memory: 37888k
        Processing time (w/o IO): 1555ms
        total: 1.64s memory: 37760k
        Processing time (w/o IO): 1461ms
        total: 1.54s memory: 38016k
        Processing time (w/o IO): 1456ms
        total: 1.54s memory: 37888k
        Processing time (w/o IO): 1447ms
        total: 1.54s memory: 38016k
        Processing time (w/o IO): 1527ms
        total: 1.62s memory: 37888k
        Processing time (w/o IO): 1570ms
        total: 1.66s memory: 37888k
        Processing time (w/o IO): 1552ms
        total: 1.64s memory: 38016k
        Processing time (w/o IO): 1485ms
        total: 1.57s memory: 38016k
        Processing time (w/o IO): 1483ms
        total: 1.57s memory: 38016k

Ruby:

        Processing time (w/o IO): 22562ms
        total: 22.78s memory: 79104k
        Processing time (w/o IO): 23253ms
        total: 23.47s memory: 79104k

Ruby:

        Processing time (w/o IO): 218813ms
        total: 219.37s memory: 184860k
        Processing time (w/o IO): 208466ms
        total: 209.08s memory: 184068k

```