@d-krushinsky 2024-10-10
@jinyus thanks for merge
looks like in main you use old ruby version, 3.2.5 I guess since it's latest version in arch packages, but there is newer version in Extra-Testing repo, not sure if it will be a good idea to install smth from Extra-Testing, but it will be nice to run benchmark in main with it (or just rerun when it comes out of Extra-Testing repo), it should be faster (probably 1.1s-1.3s) and use less memory

https://archlinux.org/packages/?name=ruby